+++ tomcat/src/main/java/nextstep/jwp/controller/AbstractController.java
@@ -1,6 +1,6 @@
 package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.HttpMethod;
+import org.apache.coyote.http11.support.HttpMethod;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
