+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -5,6 +5,7 @@ import org.apache.coyote.Processor;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.request.HttpRequestFactory;
 import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.support.ResponseProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
