+++ tomcat/src/main/java/org/apache/coyote/http11/support/HttpCookie.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.support;
 
 import java.util.Arrays;
 import java.util.Map;
