+++ tomcat/src/main/java/org/apache/coyote/http11/support/HttpStatus.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.support;
 
 public enum HttpStatus {
     OK("200 OK "),
